+++ release.ps1
@@ -1,5 +1,4 @@
-cargo publish -p dd-manifest-tree
-cargo publish -p device-driver-generation
+cargo publish -p device-driver-compiler
 cargo publish -p device-driver-macros
 cargo publish -p device-driver
 cargo publish -p device-driver-cli